+++ docs/CHANGELOG.md
@@ -7,3 +7,4 @@
   
 ## [v0.1.2]
 - quick fix for `get_sites_stations_list()`
+- hardcoded `SSTC` station to the stations list